+++ src/csvpack/writer.py
@@ -90,6 +90,18 @@ class CSVPackWriter:
 
         return self._writer.dtypes
 
+    @property
+    def num_columns(self) -> int:
+        """Get number of columns."""
+
+        return len(self.columns)
+
+    @property
+    def num_rows(self) -> int:
+        """Get number of rows read so far."""
+
+        return self._writer.num_rows
+
     def __validate_write_state(self) -> None:
         """Validate expected parameters."""